+++ pyeapi/api/spanningtree.py
@@ -34,4 +34,4 @@ import warnings
 warnings.warn("Api module spanningtree is dereciated.  Please update api "
               "calls to use stp instead")
 
-from pyeapi.api.stp import instance
+from pyeapi.api.stp import instance     # flake8: noqa